+++ packages/expo-modules-core/ios/Core/AppContext.swift
@@ -394,13 +394,12 @@ public final class AppContext: NSObject {
 
   internal func prepareRuntime() throws {
     let runtime = try runtime
-    let coreObject = try coreModuleHolder.definition.build(appContext: self)
+    let coreObject = runtime.createObject()
+
+    try coreModuleHolder.definition.decorate(object: coreObject, appContext: self)
 
     // Initialize `global.expo`.
     try runtime.initializeCoreObject(coreObject)
-
-    // Install the modules host object as the `global.expo.modules`.
-    EXJavaScriptRuntimeManager.installExpoModulesHostObject(self)
 
     // Install `global.expo.EventEmitter`.
     EXJavaScriptRuntimeManager.installEventEmitterClass(runtime)
@@ -409,6 +408,12 @@ public final class AppContext: NSObject {
     EXJavaScriptRuntimeManager.installSharedObjectClass(runtime) { [weak sharedObjectRegistry] objectId in
       sharedObjectRegistry?.delete(objectId)
     }
+
+    // Install `global.expo.NativeModule`.
+    EXJavaScriptRuntimeManager.installNativeModuleClass(runtime)
+
+    // Install the modules host object as the `global.expo.modules`.
+    EXJavaScriptRuntimeManager.installExpoModulesHostObject(self)
   }
 
   /**